Fix for when CloudFlare challenge answers are too large to be stored in an Integer.

+++ akhyou2/src/main/java/dulleh/akhyou/Utils/CloudflareSolver.java
@@ -65,7 +65,7 @@ public class CloudflareSolver {
         String function = transformFunction(doc.select("head script").first().html());
 
         Duktape duktape = Duktape.create();
-        Integer answer = Integer.valueOf(duktape.evaluate(function)) + host.length();
+        Long answer = Long.valueOf(duktape.evaluate(function)) + host.length();
         duktape.close();
 
         HttpUrl submitUrl = HttpUrl.parse(request.url().uri().getScheme() + "://" + host +"/cdn-cgi/l/chk_jschl").newBuilder()
